feat: implement camera-relative rendering for deep zoom precision and fix tick label overlap
- Change TickResult::positions from std::vector<float> to std::vector<double> for higher precision tick generation
- Add camera-relative coordinate system to 2D rendering:
  - Build orthographic projection centered at view midpoint with zero translation terms
  - Upload series data relative to per-series origin (view center) in double precision before float conversion
  - Add origin_x/origin_y to SeriesG

+++ include/spectra/axes.hpp
@@ -40,7 +40,7 @@ struct AxisLimits
 
 struct TickResult
 {
-    std::vector<float>       positions;
+    std::vector<double>      positions;
     std::vector<std::string> labels;
 };
